requirements.txt
asgiref==3.4.1
astroid==2.3.3
certifi==2021.5.30
colorama==0.4.3
Django==3.2.7
djangorestframework==3.12.4
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
pylint==2.4.4
pytz==2021.1
sqlparse==0.4.2
wincertstore==0.2
